pyproject.toml
[tool.poetry]
name = "image-to-text"
version = "0.1.0"
description = "Extract text from an image using Tesseract OCR."
authors = ["Mitch Haddadi <12067255+dfirsec@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytesseract = "^0.3.10"
pillow = "^10.4.0"
opencv-python = "^4.10.0.84"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
